refactor: extract graph and ref resolvers
