+++ .flow/tasks/fn-84-typed-graph-traversal-and-retrieval.1.md
@@ -17,7 +17,7 @@ Add the typed-edge data model: a derived `doc_edges` table with **distinct** new
 - **Stale-edge model (critical):** `doc_edges` caches a resolved `dst_doc_id`, but `markInactive()` soft-deletes (`documents.active = 0`, `adapter.ts:1038-1051`) so `ON DELETE CASCADE` won't fire. Add a **`backfillDocEdges()` projection service** — idempotent + transactional — that (re)derives edges for the **settled** document set; it is the post-migration backfill runner AND the sync-repair re-projection. `dst_doc_id` is a derived cache rebuilt by projection, not a hard referential contract. (Task .2 wires it into the sync post-upsert pass.)
 - **Persist `content_type_source`** as a new column on `documents` (migration v10) — currently sync-result-only; diagnose (.4) needs to read it from the DB.
 - Migration v10: additive/idempotent, PRAGMA-guarded, no-op `down` for table/column — follow `009-content-type-rule-fingerprint.ts`; register in `migrations/index.ts`. Schema migration stays separable from the data backfill.
-- **Backfill reuses the resolver extracted in task .9** (`src/core/graph-resolver.ts`) — do NOT re-implement resolution in SQL. Project resolved `doc_links` → `mentions`/`related_to` (`source: wikilink|markdown-link`, `confidence: parsed`). Parity test: backfilled edges == edges `getGraph` derives for the same data.
+- **Backfill reuses the resolver extracted in task .9** (`src/core/graph-resolver.ts`) — call the shared SQL helper builders instead of inlining a second resolver. Project resolved `doc_links` → `mentions`/`related_to` (`source: wikilink|markdown-link`, `confidence: parsed`). Parity test: backfilled edges == edges `getGraph` derives for the same data.
 - ADR-007 records: derived-table choice, `UNIQUE(...,source)` + read dedup, free-form `edge_type`, distinct type names; references the .9 resolver extraction.
 
 ## Investigation targets
@@ -27,7 +27,8 @@ Add the typed-edge data model: a derived `doc_edges` table with **distinct** new
 - `spec/db/schema.sql:243-266` — `doc_links` (CHECK/uniqueness style to mirror)
 - `src/store/migrations/009-content-type-rule-fingerprint.ts` + `migrations/index.ts` — template + registration
 - `src/store/sqlite/adapter.ts:1803-1926` — `setDocLinks`/read patterns to mirror
-- `src/store/sqlite/adapter.ts:2301` + `src/core/links.ts` — the resolver backfill must reuse for `getGraph` parity
+- `src/core/graph-resolver.ts` + `src/store/sqlite/adapter.ts:2301` — shared resolver helpers now used by `getGraph`; backfill must reuse them for parity
+- `src/core/ref-parser.ts` — shared ref parser/resolver extracted in task .9 for non-CLI callers
 - `src/store/types.ts:178-239,428` — `DocLinkRow/Input`, existing `GraphEdgeConfidence` (collision to avoid)
 
 **Optional:**
@@ -41,7 +42,7 @@ Add the typed-edge data model: a derived `doc_edges` table with **distinct** new
 - [ ] `setDocEdges(documentId, edges, source)` replace-by-source + typed-edge read methods that join `documents.active=1` (src+dst) and dedup by `(src,dst,edge_type)` with precedence `manual>configured>parsed>inferred`
 - [ ] `backfillDocEdges()` projection service: idempotent, transactional, runnable post-migration and as sync-repair re-projection; no reliance on `ON DELETE CASCADE`
 - [ ] `content_type_source` persisted as a `documents` column via v10
-- [ ] Backfill reuses the **task .9** resolver; **parity test** proves backfilled edges match `getGraph`-derived edges
+- [ ] Backfill reuses the **task .9** resolver helpers; **parity test** proves backfilled edges match `getGraph`-derived edges
 - [ ] Existing `linkType: wiki|markdown` APIs/schemas, `getGraph`, `gno links`/`gno backlinks` read paths unchanged on old data
 - [ ] ADR-007 records the data-model decision
 - [ ] Regression tests cover edge CRUD, replace-by-source, idempotency, backfill parity, backward compatibility
