+++ .flow/tasks/fn-84-typed-graph-traversal-and-retrieval.3.md
@@ -11,7 +11,7 @@ Build the **shared bounded-traversal core** (recursive CTE over `doc_edges` from
 
 ## Approach
 
-- **Shared core (the deliverable REST/MCP consume):** a bounded traversal function over `doc_edges` SQL — **recursive CTE BFS from a single resolved root** (resolve via the **core ref module `src/core/ref-parser.ts`** extracted in task .1, NOT `src/cli/commands/ref-parser.ts`, and NOT the global capped `getGraph` export at `adapter.ts:2301` which is wrong/unsafe for per-root traversal of isolated roots). Direction via separate inbound/outbound `UNION` arms; cycle safety via `UNION` + delimiter-wrapped `instr` (not `LIKE`); deterministic `ORDER BY (depth, edge_type, dst)` before any `LIMIT`. Reads use the `(src,dst,edge_type)` dedup from task .1.
+- **Shared core (the deliverable REST/MCP consume):** a bounded traversal function over `doc_edges` SQL — **recursive CTE BFS from a single resolved root** (resolve via the **core ref module `src/core/ref-parser.ts`** extracted in task .9, NOT `src/cli/commands/ref-parser.ts`, and NOT the global capped `getGraph` export at `adapter.ts:2301` which is wrong/unsafe for per-root traversal of isolated roots). Direction via separate inbound/outbound `UNION` arms; cycle safety via `UNION` + delimiter-wrapped `instr` (not `LIKE`); deterministic `ORDER BY (depth, edge_type, dst)` before any `LIMIT`. Reads use the `(src,dst,edge_type)` dedup from task .1.
 - **Bounded frontier mechanics (hub-node safety) — not just final caps:** a final `ORDER BY ... LIMIT` can still enumerate a huge frontier before truncation on a hub node. Enforce a **hard depth clamp**, a **per-depth frontier cap**, and a **SQL-level visited-row cap inside the recursion** (e.g. bound the recursive set / `LIMIT` within the CTE), so intermediate work stays bounded. Emit `truncated` when any cap trips. Prove with a hub-graph runtime test, not only assertions on final node/edge counts.
 - `gno graph query` registered in `wireGraphCommand` (`program.ts:2280`) — wraps the core. Add command id + format matrix entry in `src/cli/options.ts:21` `CMD`.
 - New `graph-query.schema.json` (nodes/edges with `edgeType`/`relationType`/`confidence`/`edgeSource`, depth, `truncated`; **per-node `graphHints` field** from the node's content type so non-primary hints are surfaced, not just prose) — **`schemaVersion` required** (new schema). **Register the schema name in `test/spec/schemas/validator.ts`** (it hardcodes a `schemaFiles` array — new schemas won't load otherwise). Contract test mirroring `test/spec/schemas/api-graph.test.ts`.
