+++ .flow/tasks/fn-84-typed-graph-traversal-and-retrieval.4.md
@@ -12,7 +12,7 @@ Build the **shared `diagnoseQueryTarget()` core** that explains where a named ta
 ## Approach
 
 - **Why not just `explain.ts`:** `searchHybrid` (`hybrid.ts:250`) discards raw per-source BM25/vector scores, per-stage candidate lists, absent-target info, and filter-drop reasons after RRF. Add an **opt-in trace** option to `searchHybrid` (gated — off for normal queries, **no perf regression**) that records, per stage, the candidate doc/chunk ids + **raw scores before RRF** (RRF is rank-based; magnitudes are lost after). Thread trace types through `pipeline/types.ts`; capture pre-fusion ranks in `fusion.ts`.
-- **`diagnoseQueryTarget(deps, query, target, options)`** (the shared deliverable) — `deps` mirror `searchHybrid`'s dependencies (store + ports), `options` carry model/depth-policy/filters; don't hide them in a bare `filters` arg. Resolve the target first (via the core `src/core/ref-parser.ts` from task .1), fetch its document + chunk ids, run the traced pipeline, then compare the target's chunk ids against each stage's candidate set. This is what makes `not_in_candidate_set` (retrieval miss) vs `below_cutoff` (ranking miss) decidable — reuse `buildExplainResults`/`explain*` (`explain.ts:207`) for per-result sub-scores.
+- **`diagnoseQueryTarget(deps, query, target, options)`** (the shared deliverable) — `deps` mirror `searchHybrid`'s dependencies (store + ports), `options` carry model/depth-policy/filters; don't hide them in a bare `filters` arg. Resolve the target first (via the core `src/core/ref-parser.ts` from task .9), fetch its document + chunk ids, run the traced pipeline, then compare the target's chunk ids against each stage's candidate set. This is what makes `not_in_candidate_set` (retrieval miss) vs `below_cutoff` (ranking miss) decidable — reuse `buildExplainResults`/`explain*` (`explain.ts:207`) for per-result sub-scores.
 - **`targetStatus`** (set before stage tracing): `not_found` (ref doesn't resolve), `inactive` (resolved doc not active/indexed), `no_indexed_content` (no `mirrorHash`/chunks), `filtered_out` (excluded by an active filter pre-retrieval), `diagnosed` (full stage trace produced). Only `diagnosed` runs the per-stage comparison.
 - **`filtered_out` must evaluate the FULL query filter set** the live query path applies — collection, tags, `lang`/chunk language, `exclude`, category/`contentType`, author, date — not only `matchesDocumentFilters` (`graph-retrieval.ts:58`, which covers author/contentType/category) plus `tags` via the separate tag filter. Centralize the filter evaluation so diagnose and the real query agree.
 - Per-stage payload: `present`, `rank`, `score`, `survived`, `dropReason`. Typed metadata reads `contentType`, `content_type_source` (now a persisted column via task .1/.2), categories, fingerprint match/mismatch, **`graphHints`** (the target's content-type hints, surfaced as a real field), plus chunk/line explanation.
