+++ src/mcp/tools/multi-get.ts
@@ -10,7 +10,7 @@ import type { DocumentRow, StorePort } from "../../store/types";
 import type { ToolContext } from "../server";
 
 import { decorateUriForIndex, parseUri } from "../../app/constants";
-import { parseRef } from "../../cli/commands/ref-parser";
+import { parseRef } from "../../core/ref-parser";
 import { runTool, type ToolResult } from "./index";
 
 interface MultiGetInput {
